[#21] test: DocumentServiceTest 생성과 조회 분리
